feat(simulation): add deterministic slot driver
Promote the M4 simulation spike into reusable allocdb-node test support with seeded same-slot scheduling, explicit logical-slot advancement, and restart/persist-failure helpers.

Add regression coverage for reproducible transcripts, deterministic lag/backlog behavior, and the expiration restart path, then update the simulation docs and status snapshot.
